requirements.txt
langchain
langchain_core 
numpy
python-dotenv
langchain_community
streamlit
langchain_ollama